refactor(phase4): extract utilities and config from components
date-utils.ts (new):
- Move toIsoDateTime() and getFrenchMonth() out of event-card.tsx
- Replace switch statement with Map for getFrenchMonth
- event-card.tsx now imports from ~/lib/date-utils
- event-card-datetime.test.tsx updated to import from new location

config/form-options.ts (new):
- Extract AVAILABILITY_OPTIONS from contact-form.tsx

data/contact.ts (new):
- Extract hardcoded contact info (email, location, texts) from contact-info.tsx
- contact-info.tsx now imports CONTACT_INFO from ~/data/contact

data/about.ts:
- Add ABOUT_DEFAULT_CONTENT constant
- about-content.tsx now imports from ~/data/about instead of inline string
