+++ app/components/layout/contact/contact-info.tsx
@@ -1,6 +1,7 @@
-import type * as React from "react";
-import { cn } from "~/lib/utils";
 import { Mail, MapPin } from "lucide-react";
+import type * as React from "react";
+import { CONTACT_INFO } from "~/data/contact";
+import { cn } from "~/lib/utils";
 
 /**
  * Props for the ContactInfo component
@@ -51,43 +52,40 @@ export function ContactInfo({ className, ...props }: ContactInfoProps) {
           "text-primary mb-4",
         )}
       >
-        Écris-moi
+        {CONTACT_INFO.title}
       </h2>
 
       {/* Introduction */}
       <p className={cn("font-body text-base sm:text-lg leading-relaxed", "text-primary/90 mb-2")}>
-        Contacte-moi pour échanger sur ton parcours, tes besoins et découvrir comment je peux te
-        soutenir à chaque étape. Chaque expérience est unique, créons-la ensemble !
+        {CONTACT_INFO.intro}
       </p>
 
       {/* Location */}
       <div className="flex items-start gap-3 mb-2">
         <MapPin className="w-5 h-5 mt-1 flex-shrink-0 text-primary" aria-hidden="true" />
-        <p className="font-body text-base sm:text-lg text-primary/90">Grande région de Montréal</p>
+        <p className="font-body text-base sm:text-lg text-primary/90">{CONTACT_INFO.location}</p>
       </div>
 
       {/* Email */}
       <div className="flex items-start gap-3 mb-4">
         <Mail className="w-5 h-5 mt-1 flex-shrink-0 text-primary" aria-hidden="true" />
         <a
-          href="mailto:pauline@paulineroussel.ca"
+          href={`mailto:${CONTACT_INFO.email}`}
           className={cn(
             "font-body text-base sm:text-lg text-primary hover:text-primary/80",
             "underline underline-offset-4 transition-all duration-200",
-            // Accessibility - focus states
             "outline-none focus-visible:ring-2 focus-visible:ring-primary/50 focus-visible:ring-offset-2",
-            // Touch target - pregnancy-safe minimum 44px
             "min-h-[44px] flex items-center",
           )}
           aria-label="Envoyer un courriel à Pauline Roussel"
         >
-          pauline@paulineroussel.ca
+          {CONTACT_INFO.email}
         </a>
       </div>
 
       {/* Personal message */}
       <p className={cn("font-heading text-2xl sm:text-3xl italic", "text-primary mt-4")}>
-        J'ai hâte de faire ta rencontre
+        {CONTACT_INFO.closingMessage}
       </p>
     </div>
   );
