+++ app/config/form-options.ts
@@ -1,0 +1,10 @@
+/**
+ * Form option constants for the contact form.
+ */
+
+export const AVAILABILITY_OPTIONS = [
+  { value: "", label: "Flexible" },
+  { value: "morning", label: "Matin (9h-12h)" },
+  { value: "afternoon", label: "Après-midi (12h-17h)" },
+  { value: "evening", label: "Soir (17h-20h)" },
+] as const;
