+++ app/data/contact.ts
@@ -1,0 +1,13 @@
+/**
+ * Static contact information for Pauline Roussel.
+ * Update this file when contact details change.
+ */
+
+export const CONTACT_INFO = {
+  email: "pauline@paulineroussel.ca",
+  location: "Grande région de Montréal",
+  title: "Écris-moi",
+  intro:
+    "Contacte-moi pour échanger sur ton parcours, tes besoins et découvrir comment je peux te soutenir à chaque étape. Chaque expérience est unique, créons-la ensemble !",
+  closingMessage: "J'ai hâte de faire ta rencontre",
+} as const;
